Unit tests more readable and closer to Perl code.
Some tests have been made to look more similar to the Perl code which
generates the data. This makes use of the fact that the "exected"
results are now read in by using "compile" instead of simpy reading the
expected files as text. Thus, we can use references properly. In Python
this makes not such a huge difference than in Perl. It primarily makes
the tests easier to follow.