Removed obsolete "freeze" tests.
In the existing cases where the "freeze" result was separated from the
"normal" result. This was done for tests working with floating point
numbers. Python would consider the values to be equal anyway (due to
floating point imprecisions). As such, separating the tests is
unnecessary. I removed those to keep working with the tests a bit
easier.