Much more detailed tests.
Unit tests are now generated dynamically.  TestClass is generated for
each architecture, then unit-tests are auto-generated from files
generated by the "test.pl" script.

This increases the number of tests dramatically (from around 40 to
around 2000 at the time of this writing). The advantage of this is that
test-output is much more detailed, and tests are no longer prone to
failures caused by non-deterministic data-structures like dictionaries.